requirements.txt
asgiref==3.2.10
attrs==20.2.0
Django==3.1.2
djangorestframework==3.12.1
factory-boy==3.1.0
Faker==4.9.0
importlib-metadata==2.0.0
iniconfig==1.0.1
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.1
zipp==3.3.0
